move cli util functions into a separate folder
