+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod core;
 pub mod network;
 pub mod node;
 pub mod storage;
+pub mod utils;
 
 mod tests;
